refactor: 불필요한 Computer 객체를 삭제하고, 상수는 Referee 객체로 이동
