+++ src/main/kotlin/baseball/domain/Referee.kt
@@ -1,8 +1,4 @@
-package baseball.domain.computer
-
-import baseball.domain.Score
-import baseball.domain.computer.Computer.Companion.RESET_NUMBER
-import baseball.domain.computer.Computer.Companion.THREE_STRIKE
+package baseball.domain
 
 class Referee {
     private val score = Score(RESET_NUMBER, RESET_NUMBER)
@@ -29,5 +25,13 @@ class Referee {
 
     fun isThreeStrike(): Boolean {
         return score.strike == THREE_STRIKE
+    }
+
+    companion object {
+        const val MIN_RANDOM_NUMBER = 1
+        const val MAX_RANDOM_NUMBER = 9
+        const val THREE_STRIKE = 3
+        const val RESET_NUMBER = 0
+        const val MAX_COM_NUMBER_SIZE = 3
     }
 }
