+++ src/test/kotlin/study/BaseballTest.kt
@@ -2,8 +2,7 @@ package study
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-import baseball.domain.computer.Referee
+import baseball.domain.Referee
 
 class BaseballTest {
